Finished implementing the main recycler view list of parks.
- Created the Adapter
- Created an fragment for each list item.
- Other smaller changes.

+++ app/src/main/java/com/example/parcmarc/Park.kt
@@ -5,10 +5,10 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "park")
-class Park (
-        @ColumnInfo var name: String,
-        @ColumnInfo var latitude: Float,
-        @ColumnInfo var longitude: Float,
+class Park(
+    @ColumnInfo var name: String,
+    @ColumnInfo var latitude: Double,
+    @ColumnInfo var longitude: Double,
 ) {
 
     @PrimaryKey(autoGenerate = true) var id: Long = 0
